+++ go/cmd/vtctld/plugin_zktopo.go
@@ -133,7 +133,7 @@ func (ex ZkExplorer) addTabletLinks(data string, result *ZkResult) {
 	}
 
 	if port, ok := t.Portmap["vt"]; ok {
-		result.Links["status"] = template.URL(fmt.Sprintf("http://%v/debug/status", netutil.JoinHostPort(t.Hostname, port)))
+		result.Links["status"] = template.URL(fmt.Sprintf("http://%v/debug/status", netutil.JoinHostPort(t.Hostname, int32(port))))
 	}
 }
